Refactored clients
