+++ sdk/openai/inference/src/clients/azure_openai_client.rs
