+++ sdk/openai/inference/src/clients/openai_client.rs
