+++ sdk/openai/inference/src/lib.rs
@@ -3,7 +3,7 @@ mod clients;
 mod models;
 mod options;
 
-pub use clients::azure::*;
-pub use clients::non_azure::*;
+pub use clients::azure_openai_client::*;
+pub use clients::openai_client::*;
 pub use models::*;
 pub use options::*;
